+++ docs/CHANGELOG.md
@@ -22,8 +22,8 @@
 * 🌈 Encodings:
   * [`oneVPL` codecs](https://github.com/Xpra-org/xpra/pull/4886)
   * [`libde265` encoder](https://github.com/Xpra-org/xpra/issues/4464)
-  * [`libva` encoder](https://github.com/Xpra-org/xpra/issues/4628)[ and decoder](https://github.com/Xpra-org/xpra/issues/4907)
-  * [macOS `VideoToolbox` encoder](https://github.com/Xpra-org/xpra/issues/4908)
+  * [`libva` encoder](https://github.com/Xpra-org/xpra/issues/4628) [and decoder](https://github.com/Xpra-org/xpra/issues/4907)
+  * [macOS `VideoToolbox` encoder](https://github.com/Xpra-org/xpra/issues/4908) [and decoder](https://github.com/Xpra-org/xpra/issues/4930)
   * [native MS Windows DirectShow webcam backend](https://github.com/Xpra-org/xpra/issues/3336)
   * [native macOS CoreMedia webcam backend](https://github.com/Xpra-org/xpra/issues/1231)
   * [`YUV422` support with `vp9`](https://github.com/Xpra-org/xpra/issues/4928)